WIP: better visualizing comparisons

+++ rust/src/algorithms/sort/count_sort.rs
@@ -1,4 +1,4 @@
-fn array_max_value(arr: &[u32]) -> u32 {
+fn array_max_value(arr: &[usize]) -> usize {
     let mut max = arr[0];
     for &x in arr.iter() {
         if x > max {
@@ -8,17 +8,17 @@ fn array_max_value(arr: &[u32]) -> u32 {
     max
 }
 
-pub fn count_sort(arr: &mut [u32]) {
-    let k = (array_max_value(arr) + 1) as usize;
-    let mut counter = vec![0; k];
+pub fn count_sort(arr: &mut [usize]) {
+    let k = array_max_value(arr) + 1;
+    let mut counter = vec![0usize; k];
     for &x in arr.iter() {
-        counter[x as usize] += 1;
+        counter[x] += 1;
     }
     let mut i = 0usize;
     let mut j = 0usize;
     while j < k {
         if counter[j] != 0 {
-            arr[i] = j as u32;
+            arr[i] = j;
             i += 1;
             counter[j] -= 1;
         } else {
